Move KeysExporter to proper location
